requirements.txt
graphviz
PyYAML
docopt
